Rename package to io.scrollback.app
