+++ .gitignore
@@ -1,10 +1,12 @@
-.gradle
-/local.properties
-/.idea/workspace.xml
-/.idea/libraries
+*~
 .DS_Store
-/build
-=======
+
+# Keystore files
+*.jks
+
+# IDE files
+.idea
+
 # Built application files
 *.apk
 *.ap_
