+++ app/src/androidTest/java/io/scrollback/app/ApplicationTest.java
@@ -1,4 +1,4 @@
-package com.karthikb351.scrollback;
+package io.scrollback.app;
 
 import android.app.Application;
 import android.test.ApplicationTestCase;
