+++ app/src/main/java/io/scrollback/app/ScrollbackInterface.java
@@ -1,4 +1,4 @@
-package com.karthikb351.scrollback;
+package io.scrollback.app;
 
 import android.app.NotificationManager;
 import android.content.Context;
